move btc demo position
